# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Green Way Motors API
  version: 0.3.0
  description: Implemented backend contract. GreenWay_Backend_Master_Spec_v1.md is the product and API roadmap.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/v1/markets/{marketSlug}:
    get:
      tags: [Markets]
      operationId: getMarket
      parameters:
        - { name: marketSlug, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Active destination market metadata. }
        '404': { $ref: '#/components/responses/Problem' }
...
  /api/v1/admin/inventory-sources:
    get:
      tags: [Admin - Inventory Sources]
      operationId: adminListInventorySources
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/Page' }
        '403': { $ref: '#/components/responses/Problem' }
    post:
      tags: [Admin - Inventory Sources]
      operationId: adminCreateInventorySource
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/InventorySourceInput' } } }
      responses:
        '201': { $ref: '#/components/responses/Data' }
        '403': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/inventory-sources/{id}:
    parameters: [{ $ref: '#/components/parameters/UuidId' }]
    get:
      tags: [Admin - Inventory Sources]
      operationId: adminGetInventorySource
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Data' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Admin - Inventory Sources]
      operationId: adminUpdateInventorySource
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/InventorySourcePatch' } } }
      responses:
        '200': { $ref: '#/components/responses/Data' }
        '422': { $ref: '#/components/responses/Problem' }
...
  /api/v1/admin/inventory-sources/{id}/vehicles:
    parameters: [{ $ref: '#/components/parameters/UuidId' }]
    get:
      tags: [Admin - Inventory Sources]
      operationId: adminListSourceVehicles
      security: [{ bearerAuth: [] }]
      responses: { '200': { $ref: '#/components/responses/Data' } }
  /api/v1/admin/markets:
    get:
      tags: [Admin - Markets]
      operationId: adminListMarkets
      security: [{ bearerAuth: [] }]
      parameters:
        [{ $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Cursor' }]
      responses: { '200': { $ref: '#/components/responses/Page' } }
    post:
      tags: [Admin - Markets]
      operationId: adminCreateMarket
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MarketInput' } } }
      responses:
        '201': { $ref: '#/components/responses/Data' }
        '422': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/markets/{id}:
    parameters: [{ $ref: '#/components/parameters/UuidId' }]
    get:
      tags: [Admin - Markets]
      operationId: adminGetMarket
      security: [{ bearerAuth: [] }]
      responses: { '200': { $ref: '#/components/responses/Data' } }
    patch:
      tags: [Admin - Markets]
      operationId: adminUpdateMarket
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MarketPatch' } } }
      responses:
        '200': { $ref: '#/components/responses/Data' }
        '422': { $ref: '#/components/responses/Problem' }
...
components:
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: greenway_refresh_token
  responses:
...
    Data:
      description: Permission-scoped data with request ID metadata.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DataEnvelope' }
    Page:
      description: Cursor-paginated permission-scoped collection.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PageEnvelope' }
    Problem:
      description: RFC 9457 problem details.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    DataEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          oneOf:
            - { type: object, additionalProperties: true }
            - { type: array, items: { type: object, additionalProperties: true } }
        meta: { $ref: '#/components/schemas/RequestMeta' }
    PageEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { type: object, additionalProperties: true } }
        meta:
          type: object
          required: [requestId, pagination]
          properties:
            requestId: { type: string }
            pagination:
              type: object
              required: [limit, nextCursor, hasMore]
              properties:
                limit: { type: integer }
                nextCursor: { type: [string, 'null'] }
                hasMore: { type: boolean }
    RequestMeta:
      type: object
      required: [requestId]
      properties: { requestId: { type: string } }
    InventorySourceInput:
      type: object
      additionalProperties: false
      required: [companyName]
      properties:
        companyName: { type: string, minLength: 1, maxLength: 200 }
        sourceCode: { type: [string, 'null'] }
        countryId: { type: [integer, 'null'] }
        city: { type: [string, 'null'] }
        contactName: { type: [string, 'null'] }
        email: { type: [string, 'null'], format: email }
        phone: { type: [string, 'null'] }
        whatsapp: { type: [string, 'null'] }
        website: { type: [string, 'null'], format: uri }
        address: { type: [string, 'null'] }
        reference: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        paymentTerms: { type: [string, 'null'] }
        reliabilityRating: { type: [integer, 'null'], minimum: 1, maximum: 5 }
        internalNotes: { type: [string, 'null'] }
    InventorySourcePatch:
      type: object
      additionalProperties: false
      properties:
        companyName: { type: string, minLength: 1 }
        sourceCode: { type: [string, 'null'] }
        countryId: { type: [integer, 'null'] }
        city: { type: [string, 'null'] }
        contactName: { type: [string, 'null'] }
        email: { type: [string, 'null'], format: email }
        phone: { type: [string, 'null'] }
        whatsapp: { type: [string, 'null'] }
        website: { type: [string, 'null'], format: uri }
        address: { type: [string, 'null'] }
        reference: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        paymentTerms: { type: [string, 'null'] }
        reliabilityRating: { type: [integer, 'null'] }
        internalNotes: { type: [string, 'null'] }
    MarketInput:
      type: object
      additionalProperties: false
      required: [countryId, slug, currencyCode, locale]
      properties:
        countryId: { type: integer, minimum: 1 }
        slug: { type: string, pattern: '^[a-z][a-z0-9-]{1,99}$' }
        currencyCode: { type: string, pattern: '^[A-Z]{3}$' }
        locale: { type: string }
        salesEmail: { type: [string, 'null'], format: email }
        salesPhone: { type: [string, 'null'] }
        salesWhatsapp: { type: [string, 'null'] }
        seoTitle: { type: [string, 'null'] }
        seoDescription: { type: [string, 'null'] }
    MarketPatch:
      type: object
      additionalProperties: false
      properties:
        slug: { type: string }
        currencyCode: { type: string }
        locale: { type: string }
        salesEmail: { type: [string, 'null'], format: email }
        salesPhone: { type: [string, 'null'] }
        salesWhatsapp: { type: [string, 'null'] }
        seoTitle: { type: [string, 'null'] }
        seoDescription: { type: [string, 'null'] }
...
    Problem:
      type: object
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        requestId: { type: string }
...